LayoutTests/fast/dom/HTMLElement/set-and-clear-dir-attribute-expected.txt

This tests that setting the dir attribute to rtl and clearing it defaults to ltr.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(document.body).direction is "ltr"
PASS getComputedStyle(someElementInBody).direction is "ltr"
PASS document.body.dir = "rtl"; getComputedStyle(document.body).direction is "rtl"
PASS getComputedStyle(someElementInBody).direction is "rtl"
PASS someElementInBody.dir = "rtl"; getComputedStyle(someElementInBody).direction is "rtl"
PASS document.body.dir = ""; getComputedStyle(document.body).direction is "ltr"
PASS getComputedStyle(someElementInBody).direction is "rtl"
PASS someElementInBody.dir = ""; getComputedStyle(someElementInBody).direction is "ltr"
PASS someElementInBody.dir = "ltr"; getComputedStyle(someElementInBody).direction is "ltr"
PASS document.body.dir = "rtl"; getComputedStyle(document.body).direction is "rtl"
PASS someElementInBody.dir = ""; getComputedStyle(someElementInBody).direction is "rtl"
PASS successfullyParsed is true

TEST COMPLETE
